# [Geneva] Optimize metrics exporter serializer