… src/OpenTelemetry.Exporter.Geneva/Metrics/OtlpProtobuf/OtlpProtobufSerializer.cs
… src/OpenTelemetry.Exporter.Geneva/Metrics/OtlpProtobuf/ProtobufSerializerHelper.cs
… src/OpenTelemetry.Exporter.Geneva/Metrics/TlvMetricExporter.cs
… test/OpenTelemetry.Exporter.Geneva.Benchmarks/Exporter/MetricExporterBenchmarks.cs
… test/OpenTelemetry.Exporter.Geneva.Tests/OtlpProtobufMetricExporterTests.cs