@pentp 2025-03-19
I will split out the two main parts from this PR then.